loadMeta method
