+++ js/packages/common/src/actions/account.ts
@@ -7,11 +7,11 @@ import {
   TransactionInstruction,
 } from '@solana/web3.js';
 import {
-  programIds,
   SPL_ASSOCIATED_TOKEN_ACCOUNT_PROGRAM_ID,
   TOKEN_PROGRAM_ID,
   WRAPPED_SOL_MINT,
 } from '../utils/ids';
+import { programIds } from '../utils/programIds';
 import { TokenAccount } from '../models/account';
 import { cache, TokenAccountParser } from '../contexts/accounts';
