+++ js/packages/common/src/utils/index.tsx
@@ -1,5 +1,6 @@
 export * from './eventEmitter';
 export * from './ids';
+export * from './programIds';
 export * as Layout from './layout';
 export * from './notifications';
 export * from './utils';
