+++ js/packages/web/src/utils/ids.ts
@@ -1,17 +1,5 @@
 import { PublicKey } from '@solana/web3.js';
 
-export const AUCTION_PROGRAM_ID = new PublicKey(
-  'C9nHkL6BfGx9M9MyYrJqAD5hPsGJd1fHpp1uAJA6vTCn',
-);
-
-export const VAULT_PROGRAM_ID = new PublicKey(
-  '94wRaYAQdC2gYF76AUTYSugNJ3rAC4EimjAMPwM7uYry',
-);
-
-export const METAPLEX_PROGRAM_ID = new PublicKey(
-  'EPtpKdKW8qciGVd1UFyGjgbBHTbSAyvbY61h9uQGVgeu',
-);
-
 // TODO: generate key ---
 export const AR_SOL_HOLDER_ID = new PublicKey(
   'HvwC9QSAzvGXhhVrgPmauVwFWcYZhne3hVot9EbHuFTm',
